calculator.java edited online with Bitbucket

+++ src/eg/edu/alexu/csd/datastructure/calculator/cs29/calculator.java
@@ -14,7 +14,7 @@ public class calculator implements ICalculator{
 	@Override
 	public float divide(int x, int y) {
 		 
-		return x/y;
+		return (float)x/(float)y;
 	}
 
 }
